Shader files, Engine Class
The shaders are now broken down into individual files that we can access and change later on, as well as they have a method to load them in without needing to hard code in const char*.
The Engine is now built into a class rather than different functions, but overall functions the exact same. This will be done to the window as well soon.