Window Class
Window is now a proper class. There's 1-2 things that feel not totally correct but I'll keep looking around for better solutions (input callbacks is the main one).

However, everything else now feels a lot better with the Engine having a  Window member. This can still be updated and changed as I find better ways to make and include the classes as members as well as a proper graphics engine. but this is a good OO start.