Triangle Movement, different hotkeys
Added the ability to move the triangle. Need to figure out a better way to put the vertices together for a "game object" which that class needs to exist first. However, this is a good step. Will look into a better way to do that. Also need to make a graphics class that will load everything into the buffers and then the window class can draw them.