Some more updates to entity. Still really bad
Entity is now updated to draw a rectangle instead of a single triangle. Properly fixed the problems with pair so it can now properly draw the vertices member. Some updates to window to properly begin to draw things. There's still some work to be done since a decent amount of the draw calls are still hard coded, so this should change. Will continue to look at how to change this.