Moving code to better places
Added in the OpenGLHandler, which is a class that handles all OpenGL init stuff and probably some other things down the line.
Made "movement" possible in the entity class just to check that it's working.
Moved around a good amount of code in the Window class so it doesn't take care of everything related to opengl. Also made some of the static variables members since that makes more sense.
Also changed the language to use c++ 20 instead since it would be good to get some practice with it. Still need to work on making more things in the new c++ style.