Updates to Engine and Window
Made engine a unique_ptr.

First change is to the engine to use unique_ptrs for the window and the openglhandler. Not a huge change, but it's definitely a better thing to do.

Also I was experimenting with a better way to do errors, but I haven't found something that seems to solve the problem yet. The change right now uses some stringstreams which can be used later to throw exceptions and handle them.