Better abstraction
Removed vertices from Entity since that doesn't make any sense.

Added in a resize callback function. Changed deltatime to update each time the window updated. Added in the EBO so we can use it as a mesh for later. Also made all the calls for initialization better. Updated the draw function so that it will be ready for when we have entities draw in space. Also removed the random member function that just did the same thing since it didn't really add that much.

Next plan is to either draw a texture on the rectangle that we currently have or be able to draw 2 rectangles from the same buffer.