Shader Class, GLM
So clearly there are now like 400 more files to include into the project. These are all GLM files and will help when drawing down the line. Currently they aren't used too much, but are used inside Entity which we will have it save position for now.

The other biggest addition is the shader class, so now Window doesn't load everything inside of it. The window class loads a shader class which actually takes care of loading. We now also have methods of changing colors, and later can use uniforms to change positions. The biggest change is that this freed up a lot of space inside the Window class for something it didn't need to do. The shader class will slowly be extended to account for extra things that we need to do (change position, change color, and whatever else they can do).