Textures
Added a simple asset to draw with for now.

Included stb_image for texture loading. Will be incorporated into a texture class.

Window now loads a texture and displays it and has it draw in a circle. Later we can have it draw textures from the texture class, and later add textures to Entities to draw them  at their world location.

Shader now has a way to set the uniform so we can draw the vertices at a different location.

Updated the shaders so they can handle texture coordinates and draw the texture.