Texture class
Added in the texture class which handles loading textures and binding them to the gpu. Handles loading rgb and rgba images.

Moved all of the texture loading outside of the window class so that it can properly be handled inside the texture class. Now all we need is to get the id of the texture to draw it using the vertices. Later we can bind this to other things.

Also added in another texture to test out rendering 2 textures as well as loading in a rgb image.