Entity and Texture
The vertex shader now includes a mat4 which can let us od a ton of things, for now it lets us scale our texture.

Entity got updated to include a texture, as well as a non-default ctor which let's us make one with the texture. This could be updated in the future to add a texture to the entity as well as move into an ECS, but that's the distant future. For now this works.

Shader now includes a method to set mat4s, which is just used for scale right now. Later it will be updated to take care of scale/transform/rotation.

Window now has a static entity that is being used for testing. However, more of teh rendering code can start to be moved out of window and into some kind of graphics class that handles setting all the buffers and what not. Perhaps this is what the opengl handler can be used for.