Scale textures and better input
Added in the ability to scale entities with the keyboard.

In window it now takes in the scale of the entity and then scales then.

Also, input handling is now so much better, instead of clearing the map and polling before updating, it polls after and doesn't clear the map. Made inputs feel much more smooth.