Entity rotation
The beginning of rotations are happening for entities, the only thing is I need to learn how to handle quaternions a bit better (perhaps using euler angles and converting them to a quat under the hood). However, the matrix multiplication for drawing something now works properly and better (just need to keep them in the same order when moved into a better render function). This was tested before using a quaternion and was able to get it to rotate around specific vectors with the member rotation. However, it's been changed to a quat since that's what we should use later on.