Camera and Input
Biggest change is adding in the camera class, so we can now have a camera that exists in the world (still vertex space, can change to have it be in world space later) and move it around. It's its own class that pass to the window when we go to render. This way, we can figure out where in the space to render things.

Camera added to the Engine.

Entity is swapped to using a vec3 of floats which we then create a mat4 with for rendering purposes. This currently makes a bit more sense to me for now, but will still need to be changed to use quaternions under the hood.

The Input handler has been moved out of Window and probably will be updated later to account for everything the callback function handles. The key callback can probably get moved to Input as well, but we'll see if we need to do that or the current method is good enough.

Window now takes into account where the camera is in space (only one camera for now) and renders based off of where the camera is. Hence, it's update function now takes in a reference to a camera.