Error handling and Scenes
Since before errors were never handled, so upon something breaking the engine would just crash. Now, errors are properly handled so we can properly deallocate things and/or shutdown opengl when they do. This way the program remains being responsive on start/finish and doesn't cause lots of problems. Also it's definitely a smarter thing to have done. Each class that originally just exited now properly throws an error of specific types in case we need to treat some of them differently.

Other main thing is the inclusion of Scenes. I'm not quite sure how to go about implementing them, but we currently have a way to make a Scene and add it to the SceneManager, which will handle transitions to new scenes and updating them. I'm thinking each scene will have the list of Entities to update and draw, but right now the scene just contains a single Entity. I'm wondering if we should make multiple scenes which inherit from the base Scene class or do something differently. Obviously this will be some form of deserialization, but we'll get there when graphics are less poopy. However, the "game" still runs and functions exactly the same as before without having the entity object be inside the Window.