Sandbox Scene and better method of drawing
The first non-default scene is created, and we have a method of iterating through objects in the scene to draw them. It's not great and can get better, but this is how it is. However, this is a good start since each scene will have a vector of entities to draw.