Meshes
There are a ton of changes to the engine in this commit with the addition of meshes.

Engine: Draw calls are now abstracted so we reset the window, draw everything, and then swap the buffer.

Entity: Entities now contain a texture that's a shared pointer and a mesh that's a shared pointer. These eventually will be created in a mesh handler and texture handler so they can be shared. Entities now draw with their own draw function (removed from window) but don't take into account the camera.

Error: Addition of a new error type for Meshes

Mesh: This class now takes care of loading in stuff to the gpu as well as the vertices of where they should be. This probably won't change too much until the Mesh Handler gets made. The mesh contains the shader program as well.

Scenes: All scenes now have a draw step to draw each entity.

SceneManager: The scene manager now tells the current scene to draw.

Shader: Private member now instead

Texture: Function to use the texture for drawing. Private member now instead.

Window: Removed basically all the draw calls inside the window so it can be handled somewhere else. Has two functions for resetting the background and swapping the buffer to allow the draw to happen.