Fixed weird compile issue
Not sure why this didn't show up until now, but I changed from using ext.hpp to matrix_transform.hpp and the compile error of unnamed struct/union went away. I also think this is the correct header to include anyways, but it's weird that it worked until now and for some reason then broke.