Entt, Clean up rendering
So right now I've just included entt inside the project, but it doesn't currently have the ability to do anything as it's not used anywhere. This will replace the current scene entity system that I currently have, but that will be handled later.

Rendering: The big change to the actual code written inside the engine was the addition of editing how rendering worked. Previously it was drawing a frame late, but now it draws on time (at least I think so, can't check with renderdoc). The other change was decoupling the mesh and the shader since it didn't make much sense to have them combined together. for better info, go look inside the SandboxScene where 2 things are drawn. This can be expanded to even more things as long as that render process happens. I might even move some more things around but I'm not entirely sure how to do that just yet.