Starting to work with entt
Entity: This is completely removed as now entt will take place of the entity class, since an ECS is better.

Mesh: Removed the shader program from the mesh since it wasn't using it in the first place, so it doesn't need to be there.

SandboxScene: Lots of changes to using entt so we can iterate through the objects since every scene will take care of what's in each scene. Maybe this will change in the future, but this seems good so far for now.

Scene: Base scene class no longer needs texture, and the virtual functions are set to be pure virtual now. Also the set camera function was removed, and the registry takes care of the entity_list.

Scene manager doesn't need something to return the current camera or current scene (for now, may need them back later but currently they are bloat)

Shader: This is being used to test entt so I added in a default ctor. This will probably be handled by the manager but for now it at least exists.

Removed a lot of bloat inside window since there were tons of includes that didn't need to be there.