It works!
ECS now works and draws properly with at least 2 entities. Granted now it only shows one, but that's because the next task is going to be using shared pointers for the Mesh, Shader, and Texture so they don't have to be created each time. However, the render loop is now put together. The behavior from before is gone, but that can be added back inside the update loop.