ShaderManager
FragShader: Ability to render the mesh in colors or as a texture based on a uniform int

SandboxScene: Made some adjustments so there exists a Material Class (which will get moved later). Consists of a shader, mesh, and texture. Made it so the entities load in a material instead.

Shader: Renamed some stuff and also added in the default ctor to catch errors just in case the default gets used (which it shouldn't). Aslo added in some members so we can check if that exact shader already exists.

ShaderManager: The ShaderManager is a static class that exists that gets created right at runtime. What it does is create and manage the shaders of entities so we only have to make them once and reuse. There exists a method to construct a shader with a path. This will be done similarly for Texture and Mesh.

Texture: Renaming and unbinding, as well as a default ctor that throws errors if used.