Texture Manager
Texture: The mode and path are now private members so we can compare (this will change later). Also the ability to unbind a texture so it's properly handled and duplicates are not created.

TextureManager: Very similar to the ShaderManager but instead for textures. This can probably change later since it's not super great right now (if I learn another method for doing it). However, it works just the same as before.