Camera updates
Camera: The camera now follows the mouse and you can fly around and look at the objects in space. This isn't perfect just yet, but it's definitely better than before. It now contains 2 vectors inside instead of just 1: the mouse and rotation vector where the mouse takes care of previous mouse positions and the rotation vector is the pitch/yaw of the camera.

Everything else: Windows now have a mouse vector attached to them where it tells us where the mouse of that window is. Each thing that updates before now takes into account the mouse and passes it around until it ends up at the camera.

The whole project may need some restructuring since things now feel a little weird having to include a vector for almost everything to update, but it seems a little better than before.