Starting on better method of loading meshes/objects
NewMesh and NewShader will replace the Shader/Mesh/Texture class. This might be renamed to Material later but for now, this is what it is. Now sure how I'll incorporate this into the ECS to have shared things (textures/meshes) since I imagine some things will share between them.

The only thing that's not working right now is the vertices since the color isn't supposed to be in there.