Teapot and Assimp
I think I fixed the problems with assimp and the teapot for drawing, but I can't currently see the teapot. I think this is because of how I have drawing setup since it doesn't have any textures and there's no color inside the shaders. I'll try and figure this out later.