Updates to drawing
So now the draw can detect if the mesh has any textures, and if it does, apply the texture, otherwise for now it just colors them orange. Also, the improvement from the tutorial of storing the textures that we load was applied and it made it a lot faster, however I think there's something still wrong here since the backpack doesn't draw 100% correctly, so maybe this is something with the meshes or it's something with the textures. Either way, going to need some stuff to handle memory for them since right now the only way to draw is to leak memory.

However, the Capsule object now draws correctly which is great.

Perhaps everything should be modeled in a program or maybe we can just use a texture and apply it to meshes, that's something that we'll have to see.