Removal of bloat assets, extra customization
Mesh: I don't think there ever needs to be a non-1b1 mesh, but just in case I made a few changes to step in the right direction. Things are static now so in case a copy ever gets created it's not the worst thing in the world

Texture: Ability to load textures in with types, either tiled or a single asset

Transform: A bunch of other methods that make sense: action methods like translate add while setters just set the thing directly.

Sandbox: Testing of some more of the transform things. Also moved the camera creation into each scene (this needs to be done per scene)

ResourceManager: Texture types

Camera: Updated defaults (needs non-default ctor)

Window: Fullscreen adjustments. Some of this might need to be moved into a member function so that it's just a callback from the key, although checking each frame if the key is pressed is not an expensive calculation