Huge Progress
Drawing is now handled properly, where the z_order is actually taken into account. This will almost definitely update as I figure out a better method to sort the containers to draw.

Mouse: The mouse is now handled so we can left click, and it currently draws an asset at the proper location in world space.

Camera: There are some updates to the camera to do some extra world conversion for the mouse. So to go from camera space to world space.

Input Manager: Now takes into account the mouse

Renderer: Now creates a multimap so we can iterate and draw in the correct order

Window: Now acknowledges that there is a mouse