Updates to collision, tags, and a few other things
Delegators for behaviors and other things. Allows us to communicate with objects with different signal types, as well as have functions that interact with the current registry and another object if needbe.

Also all of the gui stuff got updated so that it can use normal colliders now, and if it needs the mouse that's pretty easy.

RigidBodies got removed.

Tags is now the home for all tags, including the RenderTags.

Stuff added to ignore the warnings that stb_image causes.